Fork v0.35.0 and fix undefined enum value issue
Behavior before the fork:

After running a resolver, if we validate an enum field and find an
invalid value, we throw an exception.

This seems sensible on the surface. However, if this happens while
running an asynchronous resolver, the exception is throw in the
secondary thread, after it `deliver!`s the promise. At that point, we
cannot do anything with this error: the promise has already been
delivered, so we cannot deliver a `(with-error ...)` value.

The result is that the secondary thread swallows the exception, and the
main thread hangs forever, because we never finished handling the
callback.

Solution:

When we find an undefined enum value, we return a `selector-error`
instead of throwing an exception, just like we do when we have a
coercion failure, for example.

More discussion of the issue is in the PR we opened to the upstream:
https://github.com/walmartlabs/lacinia/pull/309